v4.4.32

+++ Examples/src/java/io/agora/rtc/example/test/stress/StressSendPcmH264Test.java
@@ -43,7 +43,7 @@ public class StressSendPcmH264Test extends AgoraTest {
             final int threadId = i;
             testTaskExecutorService.execute(() -> {
                 while (checkTestTime()) {
-                    int t1 = random.nextInt(ArgsConfig.sleepTime - ArgsConfig.timeForStressLeave) + 1;
+                    int t1 = random.nextInt((int) (ArgsConfig.sleepTime - ArgsConfig.timeForStressLeave)) + 1;
                     String channel = ArgsConfig.connectionCount == 1 ? ArgsConfig.channelId
                             : ArgsConfig.channelId + threadId;
                     String connUserId = ArgsConfig.userId.equals("0") ? ArgsConfig.userId
@@ -51,7 +51,7 @@ public class StressSendPcmH264Test extends AgoraTest {
                     createConnectionAndTest(ccfg, channel, connUserId, TestTask.SEND_PCM_H264, t1);
 
                     try {
-                        Thread.sleep(ArgsConfig.sleepTime * 1000);
+                        Thread.sleep((long) (ArgsConfig.sleepTime * 1000));
                     } catch (InterruptedException e) {
                         e.printStackTrace();
                     }
